refactor: *
refactored project structure to add query functions
